requirements.txt
python-telegram-bot
PySocks
